feat: merged Ws into CText and DText
Any grammar part in which qtext/ctext/dtext is used is in a form
which 1. repeats 2. prepends FWS in the repeating part.
Which means any parser would have to parse for chars which
are q/c/dtext OR ws (and special handling if it hits another character
like "\r" indicating the start of a soft line breake etc.)